+++ src/website/src/pages/ec/content-types/job-person/person-list/docs/code.mdx
@@ -4,7 +4,7 @@ order: 2
 ---
 
 import Playground from '@ecl/website-components/Playground';
-import PersonListPageExample from '@ecl/ec-react-person-list/examples/Default';
+import PersonListPageExample from '@ecl/ec-react-person-list/examples/PersonListCore';
 
 <Playground
   system="ec"
